airflow_postgres_to_bigquery/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres
apache-airflow-providers-google
pandas
python-dotenv
astronomer-cosmos